CR feedback primarily for tests -- deployment and application status assertions
