+++ python/ray/serve/tests/unit/test_deployment_state.py
@@ -10,13 +10,10 @@ from ray._raylet import NodeID
 from ray.serve._private.autoscaling_state import AutoscalingStateManager
 from ray.serve._private.common import (
     RUNNING_REQUESTS_KEY,
-    CreatePlacementGroupRequest,
     DeploymentHandleSource,
     DeploymentID,
     DeploymentStatus,
     DeploymentStatusTrigger,
-    GangPlacementGroupRequest,
-    GangReservationResult,
     HandleMetricReport,
     ReplicaID,
     ReplicaMetricReport,
@@ -6516,551 +6513,6 @@ def test_pending_migration_prevents_in_transition_clear(
         f"Expected 0 PENDING_MIGRATION replicas but found {pending_migration_count}. "
         "The replica is stuck because _in_transition was incorrectly cleared."
     )
-
-
-class TestGangDeploymentStates:
-    def test_schedule_gang_placement_groups(self, mock_deployment_state_manager):
-        """Creates gangs successfully and verifies placement requests include expected bundles and strategy."""
-        captured_requests = []
-        gang_size = 2
-        num_gangs = 2
-        num_replicas_to_add = gang_size * num_gangs
-        replica_resource_dict = {"CPU": 2.0, "GPU": 1.0}
-        gang_strategy = "SPREAD"
-
-        def create_pg_fn(request: CreatePlacementGroupRequest, *args, **kwargs):
-            captured_requests.append(request)
-            return Mock()
-
-        create_dsm, _, _, _ = mock_deployment_state_manager
-        dsm: DeploymentStateManager = create_dsm(
-            create_placement_group_fn_override=create_pg_fn,
-        )
-        scheduler = dsm._deployment_scheduler
-        deployment_id = DeploymentID(name="d1", app_name="app1")
-        gang_request = GangPlacementGroupRequest(
-            deployment_id,
-            gang_size,
-            gang_strategy,
-            num_replicas_to_add,
-            replica_resource_dict=replica_resource_dict,
-        )
-
-        result = scheduler.schedule_gang_placement_groups({deployment_id: gang_request})
-
-        assert deployment_id in result
-        assert result[deployment_id].success
-        assert len(result[deployment_id].gang_pgs) == num_gangs
-        assert len(captured_requests) == num_gangs
-        for req in captured_requests:
-            assert isinstance(req, CreatePlacementGroupRequest)
-            assert req.bundles == [replica_resource_dict] * gang_size
-            assert req.strategy == gang_strategy
-
-    def test_schedule_gang_placement_groups_invalid_gang_size(
-        self, mock_deployment_state_manager
-    ):
-        """Returns failure when desired replicas cannot be evenly divided by gang size."""
-        gang_size = 3
-        num_replicas_to_add = 4
-        create_pg_fn = Mock()
-        create_dsm, _, _, _ = mock_deployment_state_manager
-        dsm: DeploymentStateManager = create_dsm(
-            create_placement_group_fn_override=create_pg_fn,
-        )
-        scheduler = dsm._deployment_scheduler
-        deployment_id = DeploymentID(name="d2", app_name="app2")
-        gang_request = GangPlacementGroupRequest(
-            deployment_id,
-            gang_size,
-            "STRICT_PACK",
-            num_replicas_to_add,
-            {"CPU": 1.0},
-        )
-
-        result = scheduler.schedule_gang_placement_groups({deployment_id: gang_request})
-
-        assert not result[deployment_id].success
-        assert "not divisible by gang_size" in result[deployment_id].error_message
-        create_pg_fn.assert_not_called()
-
-    def test_schedule_gang_placement_groups_all_pg_creation_failures(
-        self, mock_deployment_state_manager
-    ):
-        """Reports failure when every gang placement group creation attempt raises exceptions."""
-        gang_size = 2
-        num_gangs = 2
-        num_replicas_to_add = gang_size * num_gangs
-
-        def create_pg_fn(request: CreatePlacementGroupRequest, *args, **kwargs):
-            raise RuntimeError("simulated placement group creation failure")
-
-        create_dsm, _, _, _ = mock_deployment_state_manager
-        dsm: DeploymentStateManager = create_dsm(
-            create_placement_group_fn_override=create_pg_fn,
-        )
-        scheduler = dsm._deployment_scheduler
-        deployment_id = DeploymentID(name="d3", app_name="app3")
-        gang_request = GangPlacementGroupRequest(
-            deployment_id,
-            gang_size,
-            "STRICT_PACK",
-            num_replicas_to_add,
-            {"CPU": 1.0},
-        )
-
-        result = scheduler.schedule_gang_placement_groups({deployment_id: gang_request})
-
-        assert not result[deployment_id].success
-        assert (
-            "Failed to create any gang placement groups"
-            in result[deployment_id].error_message
-        )
-
-    def test_schedule_gang_placement_groups_partial_pg_creation_failures(
-        self, mock_deployment_state_manager
-    ):
-        """Keeps successful gang reservations when only a subset of placement groups fail."""
-        gang_size = 2
-        num_gangs = 2
-        num_replicas_to_add = gang_size * num_gangs
-        failed_gangs = 1
-        num_calls = 0
-
-        def create_pg_fn(request: CreatePlacementGroupRequest, *args, **kwargs):
-            nonlocal num_calls
-            num_calls += 1
-            if num_calls == 1:
-                raise RuntimeError("fail first gang only")
-            return Mock()
-
-        create_dsm, _, _, _ = mock_deployment_state_manager
-        dsm: DeploymentStateManager = create_dsm(
-            create_placement_group_fn_override=create_pg_fn,
-        )
-        scheduler = dsm._deployment_scheduler
-        deployment_id = DeploymentID(name="d4", app_name="app4")
-        gang_request = GangPlacementGroupRequest(
-            deployment_id,
-            gang_size,
-            "STRICT_PACK",
-            num_replicas_to_add,
-            {"CPU": 1.0},
-        )
-
-        result = scheduler.schedule_gang_placement_groups({deployment_id: gang_request})
-
-        assert result[deployment_id].success
-        assert len(result[deployment_id].gang_pgs) == num_gangs - failed_gangs
-
-    def test_schedule_gang_placement_groups_with_per_replica_bundles(
-        self, mock_deployment_state_manager
-    ):
-        """Flattens per-replica bundles and propagates label selectors and fallback strategies correctly."""
-        captured_requests = []
-        gang_size = 2
-        num_gangs = 4
-        num_replicas_to_add = num_gangs * gang_size
-
-        def create_pg_fn(request: CreatePlacementGroupRequest, *args, **kwargs):
-            captured_requests.append(request)
-            return Mock()
-
-        create_dsm, _, _, _ = mock_deployment_state_manager
-        dsm: DeploymentStateManager = create_dsm(
-            create_placement_group_fn_override=create_pg_fn,
-        )
-        scheduler = dsm._deployment_scheduler
-        deployment_id = DeploymentID(name="d5", app_name="app5")
-        per_replica_bundles = [{"GPU": 1.0, "CPU": 1.0}, {"CPU": 1.0}]
-        per_replica_label_selector = [{"gpu": "a100"}, {"zone": "z1"}]
-        per_replica_fallback = [{"allow_soft": True}, {"allow_soft": False}]
-        gang_request = GangPlacementGroupRequest(
-            deployment_id,
-            gang_size,
-            "STRICT_PACK",
-            num_replicas_to_add,
-            replica_resource_dict={"CPU": 1.0},
-            replica_placement_group_bundles=per_replica_bundles,
-            replica_pg_bundle_label_selector=per_replica_label_selector,
-            replica_pg_fallback_strategy=per_replica_fallback,
-        )
-
-        result = scheduler.schedule_gang_placement_groups({deployment_id: gang_request})
-
-        assert result[deployment_id].success
-        assert len(captured_requests) == num_gangs
-        expected_bundles = per_replica_bundles * gang_size
-        expected_label_selector = per_replica_label_selector * gang_size
-        expected_fallback = per_replica_fallback * gang_size
-        for req in captured_requests:
-            assert req.bundles == expected_bundles
-            assert req.bundle_label_selector == expected_label_selector
-            assert req.fallback_strategy == expected_fallback
-
-    def test_schedule_gang_placement_groups_without_per_replica_bundles_uses_resource_dict(
-        self, mock_deployment_state_manager
-    ):
-        """Uses replica resource dict for each gang bundle without optional selectors."""
-        captured_requests = []
-        gang_size = 3
-        num_gangs = 2
-        num_replicas_to_add = gang_size * num_gangs
-        replica_resource_dict = {"CPU": 2.0, "GPU": 0.5}
-
-        def create_pg_fn(request: CreatePlacementGroupRequest, *args, **kwargs):
-            captured_requests.append(request)
-            return Mock()
-
-        create_dsm, _, _, _ = mock_deployment_state_manager
-        dsm: DeploymentStateManager = create_dsm(
-            create_placement_group_fn_override=create_pg_fn,
-        )
-        scheduler = dsm._deployment_scheduler
-        deployment_id = DeploymentID(name="d6", app_name="app6")
-        gang_request = GangPlacementGroupRequest(
-            deployment_id,
-            gang_size,
-            "STRICT_PACK",
-            num_replicas_to_add,
-            replica_resource_dict=replica_resource_dict,
-        )
-
-        result = scheduler.schedule_gang_placement_groups({deployment_id: gang_request})
-
-        assert result[deployment_id].success
-        assert len(captured_requests) == num_gangs
-        for req in captured_requests:
-            assert req.bundles == [replica_resource_dict] * gang_size
-            assert req.bundle_label_selector is None
-            assert req.fallback_strategy is None
-
-    def test_schedule_gang_placement_groups_multiple_deployments(
-        self, mock_deployment_state_manager
-    ):
-        """Schedules gang placement groups for multiple deployments and returns independent results."""
-        create_pg_fn = Mock(return_value=Mock())
-        gang_size_1 = 2
-        num_gangs_1 = 2
-        num_replicas_to_add_1 = gang_size_1 * num_gangs_1
-        gang_size_2 = 3
-        num_gangs_2 = 2
-        num_replicas_to_add_2 = gang_size_2 * num_gangs_2
-        create_dsm, _, _, _ = mock_deployment_state_manager
-        dsm: DeploymentStateManager = create_dsm(
-            create_placement_group_fn_override=create_pg_fn,
-        )
-        scheduler = dsm._deployment_scheduler
-        deployment_id_1 = DeploymentID(name="d7", app_name="app7")
-        deployment_id_2 = DeploymentID(name="d8", app_name="app8")
-        gang_requests = {
-            deployment_id_1: GangPlacementGroupRequest(
-                deployment_id_1,
-                gang_size_1,
-                "STRICT_PACK",
-                num_replicas_to_add_1,
-                {"CPU": 1.0},
-            ),
-            deployment_id_2: GangPlacementGroupRequest(
-                deployment_id_2,
-                gang_size_2,
-                "STRICT_PACK",
-                num_replicas_to_add_2,
-                {"CPU": 1.0},
-            ),
-        }
-
-        result = scheduler.schedule_gang_placement_groups(gang_requests)
-
-        assert set(result.keys()) == {deployment_id_1, deployment_id_2}
-        assert result[deployment_id_1].success
-        assert result[deployment_id_2].success
-        assert len(result[deployment_id_1].gang_pgs) == num_gangs_1
-        assert len(result[deployment_id_2].gang_pgs) == num_gangs_2
-        assert create_pg_fn.call_count == num_gangs_1 + num_gangs_2
-
-
-class TestScaleDeploymentGangReplicas:
-    def test_stopping_replicas_skip_upscale(self, mock_deployment_state_manager):
-        """Skips upscale while gang replicas are stopping after startup failures."""
-        create_dsm, _, _, _ = mock_deployment_state_manager
-        dsm: DeploymentStateManager = create_dsm(
-            create_placement_group_fn_override=lambda *args, **kwargs: Mock(),
-        )
-        gang_size = 2
-        target_replicas = 2
-        deployment_id = DeploymentID(name="gang_stopping_skip", app_name="app")
-
-        info, version = deployment_info(
-            num_replicas=target_replicas,
-            version="v1",
-            gang_scheduling_config=GangSchedulingConfig(gang_size=gang_size),
-        )
-        dsm.deploy(deployment_id, info)
-        ds = dsm._deployment_states[deployment_id]
-
-        dsm.update()
-        check_counts(
-            ds, total=target_replicas, by_state=[(ReplicaState.STARTING, 2, version)]
-        )
-
-        for replica in ds._replicas.get([ReplicaState.STARTING]):
-            replica._actor.set_failed_to_start()
-
-        dsm._deployment_scheduler.schedule_gang_placement_groups = Mock(return_value={})
-        captured_upscales = {}
-        original_schedule = dsm._deployment_scheduler.schedule
-
-        def schedule_with_capture(upscales, downscales):
-            captured_upscales.update(upscales)
-            return original_schedule(upscales, downscales)
-
-        dsm._deployment_scheduler.schedule = Mock(side_effect=schedule_with_capture)
-        dsm.update()
-
-        assert captured_upscales == {}
-        dsm._deployment_scheduler.schedule_gang_placement_groups.assert_not_called()
-        check_counts(
-            ds, total=target_replicas, by_state=[(ReplicaState.STOPPING, 2, version)]
-        )
-        assert ds.curr_status_info.status == DeploymentStatus.UPDATING
-
-    def test_gang_reservation_failure_records_startup_failure(
-        self, mock_deployment_state_manager
-    ):
-        """Keeps upscale empty and records reservation failure details in deployment status."""
-        create_dsm, _, _, _ = mock_deployment_state_manager
-        dsm: DeploymentStateManager = create_dsm()
-        deployment_id = DeploymentID(name="gang_reservation_fail", app_name="app")
-        error_msg = "simulated gang placement reservation failure"
-
-        info, _ = deployment_info(
-            num_replicas=4,
-            gang_scheduling_config=GangSchedulingConfig(gang_size=2),
-        )
-        dsm.deploy(deployment_id, info)
-        ds = dsm._deployment_states[deployment_id]
-
-        dsm._deployment_scheduler.schedule_gang_placement_groups = Mock(
-            return_value={
-                deployment_id: GangReservationResult(
-                    success=False, error_message=error_msg
-                )
-            }
-        )
-        captured_upscales = {}
-        original_schedule = dsm._deployment_scheduler.schedule
-
-        def schedule_with_capture(upscales, downscales):
-            captured_upscales.update(upscales)
-            return original_schedule(upscales, downscales)
-
-        dsm._deployment_scheduler.schedule = Mock(side_effect=schedule_with_capture)
-        dsm.update()
-
-        assert captured_upscales == {}
-        check_counts(ds, total=0)
-        assert ds.curr_status_info.status == DeploymentStatus.UPDATING
-        assert "Gang scheduling failed" in ds.curr_status_info.message
-        assert error_msg in ds.curr_status_info.message
-
-    def test_successful_gang_reservation(self, mock_deployment_state_manager):
-        """Creates expected gang scheduling requests and marks all replicas as starting."""
-        create_dsm, _, _, _ = mock_deployment_state_manager
-        dsm: DeploymentStateManager = create_dsm()
-        gang_size = 2
-        num_gangs = 2
-        target_replicas = gang_size * num_gangs
-        deployment_id = DeploymentID(name="gang_success_sched", app_name="app")
-        gang_pgs = [Mock(name="pg-0"), Mock(name="pg-1")]
-
-        info, version = deployment_info(
-            num_replicas=target_replicas,
-            version="v1",
-            gang_scheduling_config=GangSchedulingConfig(gang_size=gang_size),
-        )
-        dsm.deploy(deployment_id, info)
-        ds = dsm._deployment_states[deployment_id]
-
-        dsm._deployment_scheduler.schedule_gang_placement_groups = Mock(
-            return_value={
-                deployment_id: GangReservationResult(success=True, gang_pgs=gang_pgs)
-            }
-        )
-
-        captured_upscales = {}
-        original_schedule = dsm._deployment_scheduler.schedule
-
-        def schedule_with_capture(upscales, downscales):
-            captured_upscales.update(upscales)
-            return original_schedule(upscales, downscales)
-
-        dsm._deployment_scheduler.schedule = Mock(side_effect=schedule_with_capture)
-        dsm.update()
-
-        assert deployment_id in captured_upscales
-        scheduling_requests = captured_upscales[deployment_id]
-        assert len(scheduling_requests) == target_replicas
-        assert {r.gang_placement_group for r in scheduling_requests} == set(gang_pgs)
-        assert sorted(r.gang_pg_index for r in scheduling_requests) == [0, 0, 1, 1]
-        check_counts(
-            ds,
-            total=target_replicas,
-            by_state=[(ReplicaState.STARTING, target_replicas, version)],
-        )
-        starting_replicas = ds._replicas.get([ReplicaState.STARTING])
-        assert len(starting_replicas) == target_replicas
-        gang_to_replicas = {}
-        for replica in starting_replicas:
-            gang_to_replicas.setdefault(replica.gang_context.gang_id, []).append(
-                replica
-            )
-
-        assert len(gang_to_replicas) == num_gangs
-        for gang_id, replicas in gang_to_replicas.items():
-            assert len(replicas) == gang_size
-            member_ids = {r.replica_id.unique_id for r in replicas}
-            assert sorted(r.gang_context.rank for r in replicas) == list(
-                range(gang_size)
-            )
-            for replica in replicas:
-                gang_context = replica.gang_context
-                assert gang_context.gang_id == gang_id
-                assert gang_context.world_size == gang_size
-                assert set(gang_context.member_replica_ids) == member_ids
-
-    def test_gang_sibling_cleanup_on_startup_failure(
-        self, mock_deployment_state_manager
-    ):
-        """Stops gang siblings when one member fails startup to avoid partial gangs."""
-        create_dsm, _, _, _ = mock_deployment_state_manager
-        dsm: DeploymentStateManager = create_dsm(
-            create_placement_group_fn_override=lambda *args, **kwargs: Mock(),
-        )
-        gang_size = 2
-        target_replicas = 4
-        deployment_id = DeploymentID(name="gang_sibling_cleanup", app_name="app")
-
-        info, version = deployment_info(
-            num_replicas=target_replicas,
-            version="v1",
-            gang_scheduling_config=GangSchedulingConfig(gang_size=gang_size),
-        )
-        dsm.deploy(deployment_id, info)
-        ds = dsm._deployment_states[deployment_id]
-
-        dsm.update()
-        starting_replicas = ds._replicas.get([ReplicaState.STARTING])
-        initial_context_by_replica = {
-            r.replica_id.unique_id: (
-                r.gang_context.gang_id,
-                r.gang_context.rank,
-                r.gang_context.world_size,
-                tuple(r.gang_context.member_replica_ids),
-            )
-            for r in starting_replicas
-        }
-        gang_to_replicas = {}
-        for replica in starting_replicas:
-            gang_to_replicas.setdefault(replica.gang_context.gang_id, []).append(
-                replica
-            )
-        failed_gang_id, failed_gang_members = next(iter(gang_to_replicas.items()))
-
-        failed_gang_members[0]._actor.set_failed_to_start()
-        failed_gang_members[1]._actor.set_ready()
-        dsm.update()
-
-        stopping_replicas = ds._replicas.get([ReplicaState.STOPPING])
-        starting_replicas = ds._replicas.get([ReplicaState.STARTING])
-        assert len(stopping_replicas) == gang_size
-        assert all(r.gang_context.gang_id == failed_gang_id for r in stopping_replicas)
-        assert all(r.gang_context.gang_id != failed_gang_id for r in starting_replicas)
-        surviving_gang_ids = {r.gang_context.gang_id for r in starting_replicas}
-        assert len(surviving_gang_ids) == 1
-        for replica in starting_replicas:
-            context_snapshot = initial_context_by_replica[replica.replica_id.unique_id]
-            assert context_snapshot == (
-                replica.gang_context.gang_id,
-                replica.gang_context.rank,
-                replica.gang_context.world_size,
-                tuple(replica.gang_context.member_replica_ids),
-            )
-        check_counts(
-            ds,
-            total=target_replicas,
-            by_state=[
-                (ReplicaState.STOPPING, 2, version),
-                (ReplicaState.STARTING, 2, version),
-            ],
-        )
-
-    def test_terminally_failed_deployment_skips_gang_reservation(
-        self, mock_deployment_state_manager
-    ):
-        """Does not reserve gang placement groups after deployment reaches terminal startup failure."""
-        create_dsm, _, _, _ = mock_deployment_state_manager
-        dsm: DeploymentStateManager = create_dsm()
-        deployment_id = DeploymentID(name="gang_terminal_failure", app_name="app")
-        info, _ = deployment_info(
-            num_replicas=2,
-            gang_scheduling_config=GangSchedulingConfig(gang_size=2),
-        )
-        dsm.deploy(deployment_id, info)
-        ds = dsm._deployment_states[deployment_id]
-
-        dsm._deployment_scheduler.schedule_gang_placement_groups = Mock(
-            return_value={
-                deployment_id: GangReservationResult(
-                    success=False, error_message="simulated gang reservation failure"
-                )
-            }
-        )
-
-        for _ in range(20):
-            dsm.update()
-            if ds.curr_status_info.status == DeploymentStatus.DEPLOY_FAILED:
-                break
-        assert ds.curr_status_info.status == DeploymentStatus.DEPLOY_FAILED
-
-        dsm._deployment_scheduler.schedule_gang_placement_groups.reset_mock()
-        dsm.update()
-        dsm._deployment_scheduler.schedule_gang_placement_groups.assert_not_called()
-
-    def test_healthy_after_starting_replicas_ready(self, mock_deployment_state_manager):
-        """Transitions gang deployment to healthy once all starting replicas become ready."""
-        create_dsm, _, _, _ = mock_deployment_state_manager
-        dsm: DeploymentStateManager = create_dsm(
-            create_placement_group_fn_override=lambda *args, **kwargs: Mock(),
-        )
-        gang_size = 2
-        target_replicas = 4
-        deployment_id = DeploymentID(name="gang_healthy", app_name="app")
-        info, version = deployment_info(
-            num_replicas=target_replicas,
-            version="v1",
-            gang_scheduling_config=GangSchedulingConfig(gang_size=gang_size),
-        )
-        dsm.deploy(deployment_id, info)
-        ds = dsm._deployment_states[deployment_id]
-
-        dsm.update()
-        check_counts(
-            ds,
-            total=target_replicas,
-            by_state=[(ReplicaState.STARTING, target_replicas, version)],
-        )
-
-        for replica in ds._replicas.get([ReplicaState.STARTING]):
-            replica._actor.set_ready()
-        dsm.update()
-
-        check_counts(
-            ds,
-            total=target_replicas,
-            by_state=[(ReplicaState.RUNNING, target_replicas, version)],
-        )
-        assert ds.curr_status_info.status == DeploymentStatus.HEALTHY
 
 
 if __name__ == "__main__":
